chore: refactored tests
